move maxouting from fluid to phi
