+++ paddle/phi/kernels/funcs/CMakeLists.txt
@@ -20,6 +20,7 @@ math_library(cross_entropy)
 math_library(im2col)
 math_library(vol2col)
 math_library(softmax DEPS math_function)
+math_library(maxouting)
 
 cc_library(
   phi_data_layout_transform
